Multi-threading with semaphores (still errors)
To synchronize an accounts balance I use semaphore to protect the setBalance() method. This will automatically protect the deposit() and withdraw() which use this method.

Normally it shows the right balance after running 20 threads depositing and withdrawing.

But when I enforce a delay in setBalance() the result is not correct.

Maybe if I use this method as synchronized. Then it will be solved.