Semaphore issue resolved
I only used semaphore for setBalance() because withdraw and deposit make use of this method.

But withdraw and deposit also use getBalance which was not sem-protected.

So I protected this time withdraw and deposit individually (and not setBalance() and getBalance() )

getBalance() could not be sem-protected because it "returns a value".